function и БД

+++ src/main/java/com/example/app2/UserRepo.java
@@ -2,5 +2,5 @@ package com.example.app2;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepo extends CrudRepository<Message, Integer> {
+public interface UserRepo extends CrudRepository<Message, Long> {
 }
